Title Screen Implementation
This is a prototype for the start screen of the game. Scenes are managed by a Scene Loader and will cycle in the order of build index. Saving System will be implemented once more of the base game has been built. Options are a work a in progress. Text Mesh Pro has been added for better text customization.