Started working on a basic enemy, and damage stuff
Player can now take damage. The player will also be knocked back when damage is taken, in one of three directions depending on the source. Right now it can be messed with to gain a lot of speed - This due to moving with both velocity and transform.position changes. Will later change movement to velocity